+++ packages/backend/src/scheduler/SchedulerTask.ts
@@ -1,4 +1,5 @@
 import {
+    isSchedulerCsvOptions,
     ScheduledEmailNotification,
     ScheduledSlackNotification,
 } from '@lightdash/common';
@@ -69,6 +70,7 @@ export const sendSlackNotification = async (
         dashboardUuid,
         channel,
         format,
+        options,
     } = notification;
     analytics.track({
         event: 'scheduler_job.started',
@@ -117,12 +119,16 @@ export const sendSlackNotification = async (
             });
         } else {
             const user = await userService.getSessionByUserUuid(userUuid);
+            const csvOptions = isSchedulerCsvOptions(options)
+                ? options
+                : undefined;
 
             let blocks;
             if (savedChartUuid) {
                 const csvUrl = await csvService.getCsvForChart(
                     user,
                     savedChartUuid,
+                    csvOptions,
                 );
                 blocks = unfurlChartCsvResults(
                     details.name,
@@ -134,6 +140,7 @@ export const sendSlackNotification = async (
                 const csvUrls = await csvService.getCsvsForDashboard(
                     user,
                     dashboardUuid,
+                    csvOptions,
                 );
                 blocks = unfurlDashboardCsvResults(
                     details.name,
@@ -198,6 +205,7 @@ export const sendEmailNotification = async (
         recipient,
         name: schedulerName,
         format,
+        options,
     } = notification;
     analytics.track({
         event: 'scheduler_job.started',
@@ -211,6 +219,7 @@ export const sendEmailNotification = async (
             type: 'email',
         },
     });
+
     try {
         const { url, details, pageType, organizationUuid } =
             await getChartOrDashboard(savedChartUuid, dashboardUuid);
@@ -237,10 +246,15 @@ export const sendEmailNotification = async (
         } else {
             const user = await userService.getSessionByUserUuid(userUuid);
 
+            const csvOptions = isSchedulerCsvOptions(options)
+                ? options
+                : undefined;
+
             if (savedChartUuid) {
                 const csvUrl = await csvService.getCsvForChart(
                     user,
                     savedChartUuid,
+                    csvOptions,
                 );
 
                 emailClient.sendChartCsvNotificationEmail(
@@ -255,6 +269,7 @@ export const sendEmailNotification = async (
                 const csvUrls = await csvService.getCsvsForDashboard(
                     user,
                     dashboardUuid,
+                    csvOptions,
                 );
                 emailClient.sendDashboardCsvNotificationEmail(
                     recipient,