+++ packages/frontend/vite.config.ts
@@ -40,6 +40,7 @@ export default defineConfig({
                 manualChunks: mapManualChunks({
                     '@blueprintjs/icons': 'blueprint-icons-vendor',
                     '@blueprintjs/': 'blueprint-vendor',
+                    '@mantine': 'mantine-vendor',
                     'highlight.js': 'highlight-vendor',
                     echarts: 'echarts-vendor',
                     '@mapbox/': 'mapbox-vendor',